Move GitHub setup from individual subprojects to root directory
This includes the pull request template and GitHub Actions workflows

For workflows each of the files is moved. GitHub Actions does not 
support subfolders for workflows so each file is also renamed with the
subproject prefix separated by _.

For the pull request template we just move one of the files and delete
the others.