Fix static analysis errors in importers
Ensure that search index always gets an integer as error code.

Strip all implementation from SAXO service. It previously has references
to class constants and the like from parent code that has been removed 
when it was deprecated.

Remove most of the the entire implementation to bring it inline with 
other deprecated vendors.