Use matrix instead of separate workflows for testing PHP versions
There does not seem to be any meaningful difference between the two
workflows so we might as well consolidate them.

PHP CS Fixer will error on PHP versions newer than 8.1 - perhaps
because of the configuration in composer.json. Use 
PHP_CS_FIXER_IGNORE_ENV=1 to ignore these errors. It should not be the
responsibility of this to to verify this anyway.