Add liip/monitor-bundle to add support for health checks
This seems to be the most popular package for handling health checks
with support for many of the services that we use e.g. Redis, RabbitMQ
etc.

Health checks are important for supporting external monitoring.

This is the result of running composer require liip/monitor-bundle and
executing the related recipe.