Replace ElasticSearch with OpenSearch in Docker Compose setup
Since this is the service we are using in production we should use it in
development as well.

Version 2 is the specificity that we use in production.

Compared to ElasticSearch the most important part of the OpenSearch 
setup is disabling the security plugin. This allows us to access the
API through HTTP in development and without providing a 
username/password.

OpenSearch has no Xpack so we can remove this config.

We can also drop access to port 9300 since this is not used by us.