Try to set .env variables based on actual environment variables
We have seen for some environment variables that even though they are
defined in the environment they are not picked up by the PHP code for
whatever reason.

To try to add some visibility into the process we use Symfony support
for using environment variables in .env files with fallback to the
current default values.

By doing this we should have visibility into the actual values of the
variables when checking the contents of .env.local.php or by running
symfony console debug:dotenv.