Bump RabbitMQ disk size to 1536 MB
Platform.sh sends out warnings with 20% and 10% disk remaining.
RabbitMQ has a low disk watermark at 100MB after which processing
stops and cron jobs stall.

Increase disk size to a value which will allow us to get both the 20%
and 10% warning so we have some time to react to address this.