Ensure vendors use UTC timezone for sources
When the image validator service tries to determine if a cover image
has been updated it tries to compare last-modified values for source
and image.

The default image validator will use UTC by default for images but
sources are created with whatever timezone is configured in the
environment. If these do not match then images will be considered
updated even though they are not.

To ensure that sources are generated with the same timezone no matter
what has been configured in the environment we hardcode source dates
to also use UTC.

An alternate approach would be to make it a configurable variable
but that would be a more complex change.