Update agency/profile usage
The current implementation would search using the default agency and
the fallback profile if no agency and profile is set.

This is problematic because messages by default will not have either
set. In these cases the system should search using both default
agency and profile. This ensures that the agency/profile combination
DK-775100/coverservice is used which will yield the widest possible
search. This is required for us to map ids from sources (either PID,
ISBN or Faust) to the remaining id types. This will make covers
searchable from all interfaces.

Update the logic for selecting agency and profile so that the default
agency and profile is used by default.