Rearrange loading schedule for vendors
Based on experience we want to prioritize some vendors which do not
run for very long and/or have important updates while others run for
very long and generate many updates.

Consequently most importantly we greatly increase the schedule for
the upload service during work days and prioritize bogportalen over
Publizon and eBookCentral.