Do not set timezone specifically for last modified dates
We need timezones to be the same across all elements of the
application.

Doctrine requires the same timezone to be used when saving and loading
dates.
https://www.doctrine-project.org/projects/doctrine-orm/en/2.19/cookbook/working-with-datetime.html#default-timezone-gotcha

If timezones are different then the system will not be able to
identify a cover file as being unchanged because the modified time
might be different.

Consequently we need to avoid setting the timezone explicitly
whereever possible.