Include sources with 0 searches in updates
Currently we only consider sources which have a single associated
search. Searches are associated with covers in the IndexMessageHandler
but errors may occur before this stage. This might includes failure
to search the material, failure to download the file etc.

In such situations we want to retry the import process automatically.
Consequently we consider all sources with a search count equal - or
lower - than the limit (1) as updated.