Add support for processing sources which do not have an image
Sometimes you may want to process a specific source which does not
have an image relation. This might be a specific vendor/id which have
failed part of the import process. This is currently not possible
with the console command because the source repository requires an
image relation.

An add option to disable this check.

Disable it by default to retain current behavior.