Upgrade symfony/amqp-messenger to 6.3.12 for php-amqp 2 compatibility
The php-amqp extension has been upgraded from v1 to version 2 on
several platforms - Platform.sh and our Docker images. This update
introduces breaking changes which causes problems with the
symfony/amqp-messenger package.

Ack() and Nack() functions are expected to return booleans but now no
longer return any value. This causes type errors to be thrown:

Symfony\Component\Messenger\Bridge\Amqp\Transport\Connection::ack(): Return
 value must be of type bool, null returned

To address this we could either downgrade the version of the php-amqp
extension to v1 but we do not have easy control over all platforms.
Consequently we go for an upgrade to ensure compatibility instead.

symfony/amqp-messenger is fully compatible with php-amqp from 6.3.12
and onwards but there is no easy way to update all three applications
to Symfony 6.3 without causing other problems raised by our CI checks.

In an attempt to address the problem we only upgrade
symfony/amqp-messenger to 6.3.12.

We do not know if bumping individual packages will cause problems but
tests so far have shown that not to be the case.